# frontend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/attendance/shifts": {
      "get": {
        "operationId": "ShiftsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Shifts"
        ]
      },
      "post": {
        "operationId": "ShiftsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShiftDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Shifts"
        ]
      }
    },
    "/api/attendance/shifts/{id}": {
      "patch": {
        "operationId": "ShiftsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateShiftDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Shifts"
        ]
      },
      "delete": {
        "operationId": "ShiftsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Shifts"
        ]
      },
      "get": {
        "operationId": "ShiftsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Shifts"
        ]
      }
    },
...
    "/api/knowledge/documents/{id}/toggle-public": {
      "post": {
        "operationId": "KnowledgeController_togglePublic",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Knowledge"
        ]
      }
    },
    "/api/knowledge/documents/{id}": {
      "delete": {
        "operationId": "KnowledgeController_deleteDocument",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Knowledge"
        ]
      }
    },
    "/api/knowledge/documents/{id}/content": {
      "get": {
        "operationId": "KnowledgeController_getDocumentContent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Knowledge"
        ]
      }
    },
    "/api/knowledge/articles": {
      "get": {
        "operationId": "KnowledgeController_listArticles",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Knowledge"
        ]
      },
      "post": {
        "operationId": "KnowledgeController_createArticle",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveKnowledgeArticleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Knowledge"
        ]
      }
    },
    "/api/knowledge/articles/{id}": {
      "get": {
        "operationId": "KnowledgeController_getArticle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Knowledge"
        ]
      },
      "put": {
        "operationId": "KnowledgeController_updateArticle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveKnowledgeArticleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Knowledge"
        ]
      }
    },
...
    "/api/knowledge/tags": {
      "get": {
        "operationId": "KnowledgeController_listTags",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Knowledge"
        ]
      },
      "post": {
        "operationId": "KnowledgeController_createTag",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveKnowledgeTagDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Knowledge"
        ]
      }
    },
    "/api/knowledge/tags/{id}/impact": {
      "get": {
        "operationId": "KnowledgeController_getTagImpact",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Knowledge"
        ]
      }
    },
...
    "/api/knowledge/tags/{id}": {
      "put": {
        "operationId": "KnowledgeController_updateTag",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveKnowledgeTagDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Knowledge"
        ]
      }
    },
    "/api/knowledge/tags/{id}/merge": {
      "post": {
        "operationId": "KnowledgeController_mergeTag",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeKnowledgeTagDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Knowledge"
        ]
      }
    },
    "/api/knowledge/tags/{id}/enable": {
      "post": {
        "operationId": "KnowledgeController_enableTag",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Knowledge"
        ]
      }
    },
    "/api/knowledge/tags/{id}/disable": {
      "post": {
        "operationId": "KnowledgeController_disableTag",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Knowledge"
        ]
      }
    },
    "/api/knowledge/categories": {
      "get": {
        "operationId": "KnowledgeController_listCategories",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Knowledge"
        ]
      },
      "post": {
        "operationId": "KnowledgeController_createCategory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveKnowledgeCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Knowledge"
        ]
      }
    },
    "/api/knowledge/categories/{id}": {
      "put": {
        "operationId": "KnowledgeController_updateCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveKnowledgeCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Knowledge"
        ]
      }
    },
    "/api/knowledge/categories/{id}/enable": {
      "post": {
        "operationId": "KnowledgeController_enableCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Knowledge"
        ]
      }
    },
    "/api/knowledge/categories/{id}/disable": {
      "post": {
        "operationId": "KnowledgeController_disableCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Knowledge"
        ]
      }
    },
    "/api/knowledge/chat/sessions/{id}/chat-stream": {
      "get": {
        "operationId": "KnowledgeChatController_chatSse",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "content",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "KnowledgeChat"
        ]
      }
    },
...
    "/api/knowledge/chat/sessions/{id}/messages": {
      "get": {
        "operationId": "KnowledgeChatController_getMessages",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "KnowledgeChat"
        ]
      }
    },
...
  },
  "info": {
    "title": "AiSystem API",
    "description": "AiSystem 后端接口文档",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateShiftDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "班次名称",
            "example": "早班"
          },
          "on_duty_time": {
            "type": "string",
            "description": "上班时间",
            "example": "09:00"
          },
          "off_duty_time": {
            "type": "string",
            "description": "下班时间",
            "example": "18:00"
          },
          "late_threshold": {
            "type": "number",
            "description": "迟到阈值 (分钟)",
            "default": 0
          },
          "early_threshold": {
            "type": "number",
            "description": "早退阈值 (分钟)",
            "default": 0
          },
          "absenteeism_threshold": {
            "type": "number",
            "description": "旷工阈值 (分钟)",
            "default": 0
          },
          "platform_id": {
            "type": "string",
            "description": "平台 ID"
          },
          "dept_id": {
            "type": "string",
            "description": "部门 ID"
          },
          "status": {
            "type": "number",
            "description": "状态 (1: 启用, 0: 禁用)",
            "default": 1
          }
        },
        "required": [
          "name",
          "on_duty_time",
          "off_duty_time"
        ]
      },
      "UpdateShiftDto": {
        "type": "object",
        "properties": {}
      },
...
      "SaveKnowledgeArticleDto": {
        "type": "object",
        "properties": {}
      },
      "SaveKnowledgeTagDto": {
        "type": "object",
        "properties": {}
      },
      "MergeKnowledgeTagDto": {
        "type": "object",
        "properties": {}
      },
      "SaveKnowledgeCategoryDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}